+++ CHANGELOG.md
@@ -90,3 +90,4 @@
 - Bug #86: Correct key features section punctuation in multiple files for consistency (@terabytesoftw)
 - Enh #87: Add `CanBeHidden` trait with `hidden()` method and tests (@terabytesoftw)
 - Enh #88: Add `CanBeAutofocus` trait with `autofocus()` method and tests (@terabytesoftw)
+- Bug #89: Remove unused `Arrays` class helper and update `Validator`class for attribute validation (@terabytesoftw)